Update quote event
